Title: 앞에서부터 삭제하기 2, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 260305/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.cpp
@@ -28,7 +28,7 @@ int main() {
         }
         pq.pop(); // 가장 작은 숫자 제외
         int pqs = pq.size();
-        double total = 0;
+        long long total = 0;
         while(pq.size())
         {
             total += (-pq.top());
